Fix: Initial Entries and New Feed Selection Test Suites
- Issue 1 fixed:  you should search for .entrys inside .feed, but you count every children of .feed instead, even if they're not .entry elements.

-Issue 2 fixed: first, you call loadFeed with index #0 but without a callback function, which means nothing special will happen when the feed is loaded.
before it has a chance to load the feed, you populate firstFeed, which means this test depends on the previous one, since .feed is in the state where the previous test left it.
then you call loadFeed with index #1 and pass done as the callback function, so jasmine will only run the next it block when the feed is loaded.